add and export styles
