+++ components/Buttons/ButtonContainer.js
@@ -1,15 +1,6 @@
-import styled from "styled-components";
 import SaveButton from "./SaveButton";
 import CancelButton from "./CancelButton";
-
-const Container = styled.section`
-  display: flex;
-  width: 100%;
-  height: 100px;
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-`;
+import { Container } from "./styles";
 
 export default function ButtonContainer({
   boxesToAdd,
